Complete redesign of upload and editing flow
Each entry now has its own page and the table on home links directly
to these pages as opposed to the work page. The /work route,
previously overloaded, now only is responsible for displaying links to
uploaded entries for that piece and allowing users to upload new
sections. Editing only ever happens on the entry page and it is no
longer possible to edit or view fingerings without an account.